+++ src/app/models/resource.interface.ts
