+++ src/app/services/resources.service.spec.ts
@@ -73,7 +73,7 @@ describe('ResourcesService', () => {
       done();
     })
 
-    let req = httpClientMock.expectOne(environment.BACKEND_DUMMY_RESOURCES);
+    let req = httpClientMock.expectOne(environment.BACKEND_SSO_RESOURCES);
     expect(req.request.method).toEqual('GET');
 
     req.flush(responseMock);
@@ -91,7 +91,7 @@ describe('ResourcesService', () => {
       }
     });
 
-    const req = httpClientMock.expectOne(environment.BACKEND_DUMMY_RESOURCES);
+    const req = httpClientMock.expectOne(environment.BACKEND_SSO_RESOURCES);
     expect(req.request.method).toEqual('GET');
 
     req.flush(errorMessage, { status, statusText: 'Internal Server Error' });
