+++ src/app/services/resources.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from 'src/environments/environment';
-import { ResourceResponse } from '../interfaces/resource.interface';
+import { ResourceResponse } from '../models/resource.interface';
 
 @Injectable({
   providedIn: 'root'
@@ -12,9 +12,8 @@ export class ResourcesService {
   constructor(private http: HttpClient) { }
 
   getResources(): Observable<ResourceResponse> {
-
     //todo: need change the env api request.
-    return this.http.get<ResourceResponse>(`${environment.BACKEND_DUMMY_RESOURCES}`,
+    return this.http.get<ResourceResponse>(`${environment.BACKEND_SSO_RESOURSES}`,
       {
         headers: {
           'Content-Type': 'application/dummy'
