+++ src/assets/dummy/resources.json
