+++ src/environments/environment.ts
@@ -8,13 +8,12 @@ export const environment = {
   BACKEND_SSO_LOGIN_URL:'/auth/login',
   BACKEND_SSO_POST_USER:'/users/me',
   BACKEND_SSO_REGISTER_URL: '/auth/register',
-  BACKEND_SSO_RESOURSES:'/resources',
+  BACKEND_SSO_RESOURCES:'../assets/dummy/resourses.json',
+  BACKEND_SSO_SOLUTION :'../assets/dummy/challenge.json',
   BACKEND_SSO_VALIDATE_TOKEN_URL:'/tokens/validate',
   ITINERARY_ID: "clpb8t1cc000008k0cg1icvl4",
   AUTHORIZATION: 'Authorization',
   BEARER: 'Bearer ',
   BACKEND_TOKEN: '',
   pageSize: 8,
-  BACKEND_DUMMY_SOLUTION :'../assets/dummy/challenge.json',
-  BACKEND_DUMMY_RESOURCES : '../assets/dummy/resourses.json'
 };
